Update Fibonacci function to use long long int
Changed the return type of the fib function and related variables from int to long long int to support larger Fibonacci numbers. Updated printf format specifiers accordingly.

+++ C/Chapter 05 Functions and Recursion/05_09_pr3.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int fib(int a);
+long long int fib(int a);
 
 int main()
 {
@@ -9,7 +9,7 @@ int main()
     for (int i = 0; i < n; i++)
     {
 
-        printf("%d\t", fib(i));
+        printf("%lld\t", fib(i));
         if (i == (n - 1))
         {
             printf("\n");
@@ -18,7 +18,7 @@ int main()
 
     return 0;
 }
-int fib(int a)
+long long int fib(int a)
 {
     if (a == 1)
     {
